Add Space Hold to ARROW and NUM Layers
Since toggling on the Arrows layer will effectively mask the FN (KBCTL Enable) key, needed a way to enter that layer on another key.

Now when ARROW or NUM Layers are toggled, the space Key acts as space on tap, and momentarily turn on the KBCTL layer when held. This is commonly known as SpaceFN to send a space on tap and momentary layer on hold.

Note that space is a regular space on all other layers