Move Kbctl key from arrow layer
Having the space FN key on whenever the arrow keys were enabled was causing issues in some applications that required the space bar to be held. Moved the space FN to the backspace key when the arrow key layer is enabled, that way we still have access to the kbctl layer.