Update previously broken firmware - VIA
REMOVE VIA

The previously uploaded firmware was broken when enabling VIA. KB would work great for a few hours then crash and report corrupt firmware, requiring a full re-flash to get it working again.

This updated firmware removes VIA and seems to be stable.

Added bonus, it's a bit smaller 😊