Bug - do not release layer lock on MO_CAPS release
Had bug where is MO_CAPS was used to lock the EXT Layer, it would automatically unlock on exit. Now checks state on MO release and leaves the layer locked if it was locked before.

The NUM LYR is not lockable, so it is ok if it is unlocked on every release