Playing with layer lock
While trying to use layer lock keys, noticed that the key only really looks at the top most layer. However, that may not be the layer that we want to lock/unlock. Made a slightly modified version of the library to see if I could get the functionality I want.